requirements.txt
django>=4.2
openpyxl>=3.1
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
reportlab>=4.0
qrcode[pil]>=7.4
pillow>=10.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-ratelimit>=4.1
